% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcd_data_2022}
\alias{rcd_data_2022}
\alias{rcd_2022_01_28}
\alias{rcd_2022_02_28}
\alias{rcd_2022_06_27}
\alias{rcd_2022_10_13}
\title{Cached datasets from \code{random.cdisc.data}}
\format{
An object of class \code{list} of length 18.

An object of class \code{list} of length 19.

An object of class \code{list} of length 19.

An object of class \code{list} of length 20.
}
\usage{
rcd_2022_01_28

rcd_2022_02_28

rcd_2022_06_27

rcd_2022_10_13
}
\description{
Cached datasets from 2022 releases of \code{random.cdisc.data}.
}
\keyword{datasets}
