% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2022_02_28.R
\docType{data}
\name{rcd_data_2022_02_28}
\alias{rcd_data_2022_02_28}
\alias{rcd_2022_02_28_adsl}
\alias{rcd_2022_02_28_adab}
\alias{rcd_2022_02_28_adae}
\alias{rcd_2022_02_28_adaette}
\alias{rcd_2022_02_28_adcm}
\alias{rcd_2022_02_28_addv}
\alias{rcd_2022_02_28_adeg}
\alias{rcd_2022_02_28_adex}
\alias{rcd_2022_02_28_adhy}
\alias{rcd_2022_02_28_adlb}
\alias{rcd_2022_02_28_admh}
\alias{rcd_2022_02_28_adpc}
\alias{rcd_2022_02_28_adpp}
\alias{rcd_2022_02_28_adqs}
\alias{rcd_2022_02_28_adrs}
\alias{rcd_2022_02_28_adsub}
\alias{rcd_2022_02_28_adtr}
\alias{rcd_2022_02_28_adtte}
\alias{rcd_2022_02_28_advs}
\title{Cached datasets from the 2022_02_28 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 44 columns.

An object of class \code{data.frame} with 1596 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 74 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 68 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 77 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 60 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 85 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4378 rows and 57 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13134 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 62 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 54 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 54 columns.

An object of class \code{data.frame} with 2800 rows and 66 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1600 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 76 columns.
}
\usage{
rcd_2022_02_28_adsl

rcd_2022_02_28_adab

rcd_2022_02_28_adae

rcd_2022_02_28_adaette

rcd_2022_02_28_adcm

rcd_2022_02_28_addv

rcd_2022_02_28_adeg

rcd_2022_02_28_adex

rcd_2022_02_28_adhy

rcd_2022_02_28_adlb

rcd_2022_02_28_admh

rcd_2022_02_28_adpc

rcd_2022_02_28_adpp

rcd_2022_02_28_adqs

rcd_2022_02_28_adrs

rcd_2022_02_28_adsub

rcd_2022_02_28_adtr

rcd_2022_02_28_adtte

rcd_2022_02_28_advs
}
\description{
Cached datasets from the 2022_02_28 release of \code{random.cdisc.data}
}
\keyword{datasets}
