% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2022_06_27.R
\docType{data}
\name{rcd_data_2022_06_27}
\alias{rcd_data_2022_06_27}
\alias{rcd_2022_06_27_adsl}
\alias{rcd_2022_06_27_adab}
\alias{rcd_2022_06_27_adae}
\alias{rcd_2022_06_27_adaette}
\alias{rcd_2022_06_27_adcm}
\alias{rcd_2022_06_27_addv}
\alias{rcd_2022_06_27_adeg}
\alias{rcd_2022_06_27_adex}
\alias{rcd_2022_06_27_adhy}
\alias{rcd_2022_06_27_adlb}
\alias{rcd_2022_06_27_admh}
\alias{rcd_2022_06_27_adpc}
\alias{rcd_2022_06_27_adpp}
\alias{rcd_2022_06_27_adqs}
\alias{rcd_2022_06_27_adrs}
\alias{rcd_2022_06_27_adsub}
\alias{rcd_2022_06_27_adtr}
\alias{rcd_2022_06_27_adtte}
\alias{rcd_2022_06_27_advs}
\title{Cached datasets from the 2022_06_27 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 46 columns.

An object of class \code{data.frame} with 64372 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 81 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 57 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 74 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 57 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 79 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 70 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 62 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 58 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4378 rows and 59 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26268 rows and 59 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 64 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 56 columns.

An object of class \code{data.frame} with 2800 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 58 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 78 columns.
}
\usage{
rcd_2022_06_27_adsl

rcd_2022_06_27_adab

rcd_2022_06_27_adae

rcd_2022_06_27_adaette

rcd_2022_06_27_adcm

rcd_2022_06_27_addv

rcd_2022_06_27_adeg

rcd_2022_06_27_adex

rcd_2022_06_27_adhy

rcd_2022_06_27_adlb

rcd_2022_06_27_admh

rcd_2022_06_27_adpc

rcd_2022_06_27_adpp

rcd_2022_06_27_adqs

rcd_2022_06_27_adrs

rcd_2022_06_27_adsub

rcd_2022_06_27_adtr

rcd_2022_06_27_adtte

rcd_2022_06_27_advs
}
\description{
Cached datasets from the 2022_06_27 release of \code{random.cdisc.data}
}
\keyword{datasets}
