% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2022_10_13.R
\docType{data}
\name{rcd_data_2022_10_13}
\alias{rcd_data_2022_10_13}
\alias{rcd_2022_10_13_adsl}
\alias{rcd_2022_10_13_adab}
\alias{rcd_2022_10_13_adae}
\alias{rcd_2022_10_13_adaette}
\alias{rcd_2022_10_13_adcm}
\alias{rcd_2022_10_13_addv}
\alias{rcd_2022_10_13_adeg}
\alias{rcd_2022_10_13_adex}
\alias{rcd_2022_10_13_adhy}
\alias{rcd_2022_10_13_adlb}
\alias{rcd_2022_10_13_admh}
\alias{rcd_2022_10_13_adpc}
\alias{rcd_2022_10_13_adpp}
\alias{rcd_2022_10_13_adqlqc}
\alias{rcd_2022_10_13_adqs}
\alias{rcd_2022_10_13_adrs}
\alias{rcd_2022_10_13_adsub}
\alias{rcd_2022_10_13_adtr}
\alias{rcd_2022_10_13_adtte}
\alias{rcd_2022_10_13_advs}
\title{Cached datasets from the 2022_10_13 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17094 rows and 22 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 92 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 84 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 89 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 80 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 72 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 103 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 68 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18179 rows and 72 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26268 rows and 69 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116803 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 74 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 66 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 66 columns.

An object of class \code{data.frame} with 2800 rows and 77 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 68 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 88 columns.
}
\usage{
rcd_2022_10_13_adsl

rcd_2022_10_13_adab

rcd_2022_10_13_adae

rcd_2022_10_13_adaette

rcd_2022_10_13_adcm

rcd_2022_10_13_addv

rcd_2022_10_13_adeg

rcd_2022_10_13_adex

rcd_2022_10_13_adhy

rcd_2022_10_13_adlb

rcd_2022_10_13_admh

rcd_2022_10_13_adpc

rcd_2022_10_13_adpp

rcd_2022_10_13_adqlqc

rcd_2022_10_13_adqs

rcd_2022_10_13_adrs

rcd_2022_10_13_adsub

rcd_2022_10_13_adtr

rcd_2022_10_13_adtte

rcd_2022_10_13_advs
}
\description{
Cached datasets from the 2022_10_13 release of \code{random.cdisc.data}
}
\keyword{datasets}
