% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2023_07_07.R
\docType{data}
\name{rcd_data_2023_07_07}
\alias{rcd_data_2023_07_07}
\alias{rcd_2023_07_07_adsl}
\alias{rcd_2023_07_07_adab}
\alias{rcd_2023_07_07_adae}
\alias{rcd_2023_07_07_adaette}
\alias{rcd_2023_07_07_adcm}
\alias{rcd_2023_07_07_addv}
\alias{rcd_2023_07_07_adeg}
\alias{rcd_2023_07_07_adex}
\alias{rcd_2023_07_07_adhy}
\alias{rcd_2023_07_07_adlb}
\alias{rcd_2023_07_07_admh}
\alias{rcd_2023_07_07_adpc}
\alias{rcd_2023_07_07_adpp}
\alias{rcd_2023_07_07_adqlqc}
\alias{rcd_2023_07_07_adqs}
\alias{rcd_2023_07_07_adrs}
\alias{rcd_2023_07_07_adsub}
\alias{rcd_2023_07_07_adtr}
\alias{rcd_2023_07_07_adtte}
\alias{rcd_2023_07_07_advs}
\title{Cached datasets from the 2023_07_07 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6916 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 92 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 66 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 83 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 119 rows and 66 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 88 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 79 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 71 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 102 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6640 rows and 72 columns.

An object of class \code{data.frame} with 26268 rows and 68 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116803 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 73 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 65 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 65 columns.

An object of class \code{data.frame} with 2800 rows and 76 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116803 rows and 50 columns.
}
\usage{
rcd_2023_07_07_adsl

rcd_2023_07_07_adab

rcd_2023_07_07_adae

rcd_2023_07_07_adaette

rcd_2023_07_07_adcm

rcd_2023_07_07_addv

rcd_2023_07_07_adeg

rcd_2023_07_07_adex

rcd_2023_07_07_adhy

rcd_2023_07_07_adlb

rcd_2023_07_07_admh

rcd_2023_07_07_adpc

rcd_2023_07_07_adpp

rcd_2023_07_07_adqlqc

rcd_2023_07_07_adqs

rcd_2023_07_07_adrs

rcd_2023_07_07_adsub

rcd_2023_07_07_adtr

rcd_2023_07_07_adtte

rcd_2023_07_07_advs

rcd_2023_07_07_adqlqc
}
\description{
Cached datasets from the 2023_07_07 release of \code{random.cdisc.data}
}
\keyword{datasets}
