% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcd_data_2021}
\alias{rcd_data_2021}
\alias{rcd_2021_03_22}
\alias{rcd_2021_05_05}
\alias{rcd_2021_07_07}
\title{Cached datasets from \code{random.cdisc.data}}
\format{
An object of class \code{list} of length 15.

An object of class \code{list} of length 16.

An object of class \code{list} of length 16.
}
\usage{
rcd_2021_03_22

rcd_2021_05_05

rcd_2021_07_07
}
\description{
Cached datasets from 2021 releases of \code{random.cdisc.data}.
}
\keyword{datasets}
