% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2023_03_17.R
\docType{data}
\name{rcd_data_2023_03_17}
\alias{rcd_data_2023_03_17}
\alias{rcd_2023_03_17_adsl}
\alias{rcd_2023_03_17_adab}
\alias{rcd_2023_03_17_adae}
\alias{rcd_2023_03_17_adaette}
\alias{rcd_2023_03_17_adcm}
\alias{rcd_2023_03_17_addv}
\alias{rcd_2023_03_17_adeg}
\alias{rcd_2023_03_17_adex}
\alias{rcd_2023_03_17_adhy}
\alias{rcd_2023_03_17_adlb}
\alias{rcd_2023_03_17_admh}
\alias{rcd_2023_03_17_adpc}
\alias{rcd_2023_03_17_adpp}
\alias{rcd_2023_03_17_adqlqc}
\alias{rcd_2023_03_17_adqs}
\alias{rcd_2023_03_17_adrs}
\alias{rcd_2023_03_17_adsub}
\alias{rcd_2023_03_17_adtr}
\alias{rcd_2023_03_17_adtte}
\alias{rcd_2023_03_17_advs}
\title{Cached datasets from the 2023_03_17 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58520 rows and 22 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 93 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 84 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 67 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 89 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 80 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 72 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 103 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 68 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6640 rows and 72 columns.

An object of class \code{data.frame} with 26268 rows and 69 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116803 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 74 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 66 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 66 columns.

An object of class \code{data.frame} with 2800 rows and 77 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 68 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 88 columns.
}
\usage{
rcd_2023_03_17_adsl

rcd_2023_03_17_adab

rcd_2023_03_17_adae

rcd_2023_03_17_adaette

rcd_2023_03_17_adcm

rcd_2023_03_17_addv

rcd_2023_03_17_adeg

rcd_2023_03_17_adex

rcd_2023_03_17_adhy

rcd_2023_03_17_adlb

rcd_2023_03_17_admh

rcd_2023_03_17_adpc

rcd_2023_03_17_adpp

rcd_2023_03_17_adqlqc

rcd_2023_03_17_adqs

rcd_2023_03_17_adrs

rcd_2023_03_17_adsub

rcd_2023_03_17_adtr

rcd_2023_03_17_adtte

rcd_2023_03_17_advs
}
\description{
Cached datasets from the 2023_03_17 release of \code{random.cdisc.data}
}
\keyword{datasets}
